skills/seo-ops/requirements.txt
# Core dependencies
requests>=2.28.0

# Google Search Console API
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
